Fix phantom delegate trying to use its referenced gatt instance
Prior to the change, when one disconnect a connected device in Android,
Device instance inserts a delegate to GattCallback instance to cache
discovered Services. However, when one disconnects such device, the
delegate is not removed. So upon next device connection, that delegate
will attempt to reference a disposed Gatt instance and cause a crash.

To reproduce this bug, you can try the following sequence:

1) Scan for Devices
2) Connect to Device A
3) Discovers Service of Device A
4) Dsiconnect Device A
5) Connect to Device A again
6) Discovers Service of Device A
7) Crash